[stagger] Spin off stagger subsystem as a mixin.
