+++ source_files/QT_engine.pri
@@ -30,6 +30,7 @@ HEADERS += engine/class_modules/apl/mage.hpp
 HEADERS += engine/class_modules/apl/warlock.hpp
 HEADERS += engine/class_modules/class_module.hpp
 HEADERS += engine/class_modules/monk/sc_monk.hpp
+HEADERS += engine/class_modules/monk/sc_stagger.hpp
 HEADERS += engine/class_modules/paladin/sc_paladin.hpp
 HEADERS += engine/class_modules/priest/sc_priest.hpp
 HEADERS += engine/class_modules/warlock/sc_warlock.hpp
@@ -226,6 +227,7 @@ SOURCES += engine/class_modules/apl/mage.cpp
 SOURCES += engine/class_modules/apl/warlock.cpp
 SOURCES += engine/class_modules/monk/sc_monk.cpp
 SOURCES += engine/class_modules/monk/sc_monk_pets.cpp
+SOURCES += engine/class_modules/monk/sc_stagger.cpp
 SOURCES += engine/class_modules/paladin/sc_paladin.cpp
 SOURCES += engine/class_modules/paladin/sc_paladin_holy.cpp
 SOURCES += engine/class_modules/paladin/sc_paladin_protection.cpp
