+++ source_files/VS_engine.props
@@ -34,6 +34,7 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\class_modules\apl\warlock.hpp" />
 		<ClInclude Include="..\engine\class_modules\class_module.hpp" />
 		<ClInclude Include="..\engine\class_modules\monk\sc_monk.hpp" />
+		<ClInclude Include="..\engine\class_modules\monk\sc_stagger.hpp" />
 		<ClInclude Include="..\engine\class_modules\paladin\sc_paladin.hpp" />
 		<ClInclude Include="..\engine\class_modules\priest\sc_priest.hpp" />
 		<ClInclude Include="..\engine\class_modules\warlock\sc_warlock.hpp" />
@@ -229,6 +230,7 @@ To change the list of source files run synchronize.py
 		<ClCompile Include="..\engine\class_modules\apl\warlock.cpp" />
 		<ClCompile Include="..\engine\class_modules\monk\sc_monk.cpp" />
 		<ClCompile Include="..\engine\class_modules\monk\sc_monk_pets.cpp" />
+		<ClCompile Include="..\engine\class_modules\monk\sc_stagger.cpp" />
 		<ClCompile Include="..\engine\class_modules\paladin\sc_paladin.cpp" />
 		<ClCompile Include="..\engine\class_modules\paladin\sc_paladin_holy.cpp" />
 		<ClCompile Include="..\engine\class_modules\paladin\sc_paladin_protection.cpp" />
