+++ source_files/engine_make
@@ -31,6 +31,7 @@ SRC += \
     class_modules$(PATHSEP)apl$(PATHSEP)warlock.cpp \
     class_modules$(PATHSEP)monk$(PATHSEP)sc_monk.cpp \
     class_modules$(PATHSEP)monk$(PATHSEP)sc_monk_pets.cpp \
+    class_modules$(PATHSEP)monk$(PATHSEP)sc_stagger.cpp \
     class_modules$(PATHSEP)paladin$(PATHSEP)sc_paladin.cpp \
     class_modules$(PATHSEP)paladin$(PATHSEP)sc_paladin_holy.cpp \
     class_modules$(PATHSEP)paladin$(PATHSEP)sc_paladin_protection.cpp \
